chore: move lint to dedicated package
Works separately from editor.
