+++ lib/editor/index.js
@@ -8,7 +8,7 @@ import { EditorView, keymap, tooltips, lineNumbers } from '@codemirror/view';
 import { parser as markdownParser } from '@lezer/markdown';
 import { createFeelersLanguageSupport } from '../lang/index.js';
 
-import { feelersLinter } from './lint/index.js';
+import { feelersLinter } from '../lint/index.js';
 import { lightTheme, darkTheme } from '@bpmn-io/cm-theme';
 import mitt from 'mitt';
