+++ lib/lint/cmFeelersLinter.js
