+++ lib/lint/index.js
