+++ lib/lint/rules/lintEmptyInserts.js
